Move pku Override to constructor
pku Overrides now take place in the Exporter class constructor, rather than as the first of the porter directive phases. As a result, the FormatOverride phase has been removed.

This is to prevent exporting pku getting around CanPort with their overrides (e.g. overriding species tag to one that doesn't exist in the format).